moves parsing code into parser class
