+++ alternatives/dotnet/discogs/Parser.cs
@@ -1,0 +1,134 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.IO.Compression;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Xml;
+using System.Xml.Serialization;
+
+namespace discogs
+{
+    public class Parser<T>
+            where T : IExportToCsv, new()
+    {
+        private readonly int _throttle = 1;
+        private readonly string _typeName;
+        public Parser(int throttle = 1)
+        {
+            _throttle = throttle;
+            _typeName = typeof(T).Name.Split('.')[^1];
+        }
+
+        public event EventHandler<ParseEventArgs> OnSucessfulParse;
+
+        public async Task ParseFileAsync(string fileName)
+        {
+            Dictionary<string, (string FilePath, StreamWriter FileStream)> csvStreams = GetCsvFilesFor(fileName);
+            int objectCount = 0;
+            using FileStream fileStream = new FileStream(fileName, FileMode.Open);
+            Stream readingStream = fileStream;
+            if (System.IO.Path.GetExtension(fileName).Equals(".gz", StringComparison.OrdinalIgnoreCase))
+            {
+                readingStream = new GZipStream(fileStream, CompressionMode.Decompress);
+            }
+            var settings = new XmlReaderSettings
+            {
+                ConformanceLevel = ConformanceLevel.Fragment,
+                Async = true,
+            };
+            using XmlReader reader = XmlReader.Create(readingStream, settings);
+
+            await reader.MoveToContentAsync();
+            await reader.ReadAsync();
+            while (!reader.EOF)
+            {
+                if (reader.Name == _typeName)
+                {
+                    var objectString = await reader.ReadOuterXmlAsync();
+                    // var objectString = await reader.ReadInnerXmlAsync();
+                    try
+                    {
+                        if (!string.IsNullOrEmpty(objectString))
+                        {
+                            // await WriteCsvAsync<T>($"<artist>{objectString}</artist>", csvStreams);
+                            await WriteCsvAsync(objectString, csvStreams);
+                        }
+                        else
+                        {
+                            continue;
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Error {ex} parsing node {objectString}");
+                    }
+                    // var labelId = DeserializeLabel(labelString);
+                    // Console.WriteLine($"label: {labelId}");
+                    // Console.Write('.');
+                    // _ = reader.ReadOuterXml();
+                    objectCount++;
+                    // await reader.SkipAsync();
+                    if (objectCount % _throttle == 0) OnSucessfulParse(null, new ParseEventArgs { ParseCount = objectCount });
+                    continue;
+                }
+                else
+                {
+                    await reader.ReadAsync();
+                }
+            }
+            var csvFileNames = string.Join("; ", csvStreams.Select(kvp => kvp.Value.FilePath));
+            // pbar.WriteLine("Parsing done. Writing streams.");
+            foreach (var kvp in csvStreams)
+            {
+                await kvp.Value.FileStream.FlushAsync();
+                kvp.Value.FileStream.Close();
+                await kvp.Value.FileStream.DisposeAsync();
+            }
+            // pbar.Dispose();
+            Console.WriteLine($"Found {objectCount:n0} {_typeName}s. Wrote them to {csvFileNames}.");
+        }
+
+        private static Dictionary<string, (string FilePath, StreamWriter FileStream)> GetCsvFilesFor(string xmlFile)
+        {
+            var obj = new T();
+            var dir = Path.GetDirectoryName(xmlFile);
+            IReadOnlyDictionary<string, string[]> files = obj.GetCsvExportScheme();
+            Dictionary<string, (string FilePath, StreamWriter FileStream)> csvFiles = files.ToDictionary(
+                kvp => kvp.Key,
+                kvp =>
+                {
+                    var csvFile = Path.Combine(dir, $"{kvp.Key}.csv");
+                    var stream = new StreamWriter(csvFile);
+                    stream.WriteLine(CsvExtensions.ToCsv(kvp.Value));
+                    return (csvFile, stream);
+                });
+
+            return csvFiles;
+        }
+
+        private static async Task WriteCsvAsync(
+            string objectString,
+            Dictionary<string, (string FilePath, StreamWriter FileStream)> streams)
+        {
+            var obj = Deserialize(objectString);
+            IEnumerable<(string StreamName, string[] Row)> csvExports = obj.ExportToCsv();
+            foreach (var (streamName, row) in csvExports)
+            {
+                await streams[streamName].FileStream.WriteLineAsync(CsvExtensions.ToCsv(row));
+            }
+        }
+
+        private static T Deserialize(string content)
+        {
+            using var reader = new StringReader(content);
+            XmlSerializer _labelXmlSerializer = new XmlSerializer(typeof(T));
+            var obj = (T)_labelXmlSerializer.Deserialize(reader);
+            return obj;
+        }
+    }
+
+    public class ParseEventArgs : EventArgs {
+        public int ParseCount {get;set;}
+    }
+}
